07-ejemplo-entrega-tareas/libros/requirements.txt
flask>=3.0
requests
flasgger
python-dotenv
gunicorn
